Default roles cannot be deleted

+++ app/Http/Controllers/RoleController.php
@@ -104,6 +104,10 @@ class RoleController extends Controller
     {
         $this->isAuthorized('delete', Role::class);
         
+        if($role->isLocked()) {
+            return abort(403);
+        }
+        
         if(! $role->isAssigned()) {
             $role->delete();
             session()->flash('message', 'Role deleted successfully');
